+++ include/eixx/marshal/tuple.hpp
@@ -124,7 +124,7 @@ public:
     /**
      * Decode the tuple from a binary buffer.
      */
-    tuple(const char* buf, int& idx, size_t size, const Alloc& a_alloc = Alloc());
+    tuple(const char* buf, uintptr_t& idx, size_t size, const Alloc& a_alloc = Alloc());
 
     ~tuple() {
         release();
@@ -176,13 +176,13 @@ public:
         return true;
     }
 
-    const eterm<Alloc>& operator[] (int idx) const {
-        BOOST_ASSERT(m_blob && (size_t)idx < size());
+    const eterm<Alloc>& operator[] (size_t idx) const {
+        BOOST_ASSERT(m_blob && idx < size());
         return m_blob->data()[idx];
     }
 
-    eterm<Alloc>& operator[] (int idx) {
-        BOOST_ASSERT(m_blob && (size_t)idx < size());
+    eterm<Alloc>& operator[] (size_t idx) {
+        BOOST_ASSERT(m_blob && idx < size());
         return m_blob->data()[idx];
     }
 
@@ -213,7 +213,7 @@ public:
         return result;
     }
 
-    void encode(char* buf, int& idx, size_t size) const;
+    void encode(char* buf, uintptr_t& idx, size_t size) const;
 
     bool subst(eterm<Alloc>& out, const varbind<Alloc>* binding) const;
